move svg map of country into separate component, pass in only strength scores.
