+++ src/USMap.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react'
-import {feature as topojsonFeature} from 'topojson'
+// import {feature as topojsonFeature} from 'topojson'
 import * as d3 from 'd3'
 import YearPicker from './YearPicker';
-import topoStates from './states'
+// import topoStates from './states'
+import CountryMap from './CountryMap';
 
 class USMap extends Component {
   constructor() {
@@ -11,7 +12,8 @@ class USMap extends Component {
     this.state = {
       year: 2016,
       censusData: [],
-      collegeData: []
+      collegeData: [],
+      strengthScores: []
     }
   }
   censusFormatter = row => {
@@ -46,6 +48,8 @@ class USMap extends Component {
     }
     return result
   }
+
+  // Read in data from csv files
   loadData() {
 
     const promises = []
@@ -60,8 +64,14 @@ class USMap extends Component {
         })
       })
       .then(() => this.refreshScores())
-      .then(() => this.loadMap())
   }
+  async refreshScores() {
+    const [ collegeYear, censusYear ] = this.getData()
+    await this.setState({
+      strengthScores: this.getScores(collegeYear, censusYear)
+    })
+  }
+  // retrieve census and college data for the specified year
   getData() {
     const collegeYear = this.state.collegeData.find(entry => {
       return entry.year === this.state.year || 
@@ -73,55 +83,7 @@ class USMap extends Component {
     })
     return [collegeYear, censusYear]
   }
-  loadMap() {
-
-    const geoDataObj = topojsonFeature(topoStates, topoStates.objects.states)
-
-    this.state.strengthScores.forEach(row => {
-      const states =  geoDataObj.features.filter(d => d.properties.name === row.name)
-      states.forEach(state => state.properties = row)
-    })
-    this.displayMap(geoDataObj)
-
-    // d3.json('https://cdn.jsdelivr.net/npm/us-atlas@2/us/states-10m.json')
-    //   .then(mapData => {
-    //     const geoDataObj = topojsonFeature(mapData, mapData.objects.states)
-
-    //     this.state.strengthScores.forEach(row => {
-    //       const states =  geoDataObj.features.filter(d => d.properties.name === row.name)
-    //       states.forEach(state => state.properties = row)
-    //     })
-    //     // testStateData.forEach(row => {
-    //     //   const states =  geoDataObj.features.filter(d => d.properties.name === row.name)
-    //     //   states.forEach(state => state.properties = row)
-    //     // })
-    //     this.displayMap(geoDataObj)
-    //   })
-    //   .catch(err => {
-    //     throw err
-    //   })
-  }
-  displayMap(geoDataObj) {
-    const width = 720
-    const height = 450
-    
-    const projection = d3.geoIdentity()
-                          .fitSize([width, height], geoDataObj)
-
-    const path = d3.geoPath().projection(projection)
-
-    d3.select('svg')
-          .attr('width', width)
-          .attr('height', height)
-        .selectAll('.state')
-        .data(geoDataObj.features)
-        .enter()
-          .append('path')
-          .classed('state', true)
-          .attr('d', path)
-
-    this.setColor()
-  }
+  // calculate scores based on data
   getScores(collegeYear, censusYear) {
     const strengthScores = []
     let popTotal = 0
@@ -131,6 +93,10 @@ class USMap extends Component {
     let voteTotal = 0
     collegeYear.states.forEach(state => {
       voteTotal += state.votes
+    })
+
+    this.setState({
+      popTotal, voteTotal
     })
 
     for (let i = 0; i < censusYear.states.length; i++) {
@@ -143,30 +109,6 @@ class USMap extends Component {
       })
     }
     return strengthScores
-  }
-  async setColor () {
-
-    // const colorRange = ['#c4e7fd', 'blue']
-    // const colorDomain = d3.extent(this.state.strengthScores.map(state => state.score))
-    const scale = d3.scaleLinear()
-                    .domain([0.5, 1, 4])
-                    .range(['red', '#FFFDE9', 'blue'])
-
-    d3.selectAll('.state')
-      .transition()
-      .duration(500)
-      .style('fill', d => {
-        const data = d.properties.score
-        return data ? scale(data) : '#aaa'
-      })
-  }
-  async refreshScores() {
-    const [ collegeYear, censusYear ] =  await this.getData()
-
-    await this.setState({
-      strengthScores: this.getScores(collegeYear, censusYear)
-    })
-    this.loadMap()
   }
   handleChange = async e => {
     await this.setState({
@@ -183,10 +125,7 @@ class USMap extends Component {
       <div>
         <h3>US Map</h3>
         <YearPicker handleChange={this.handleChange} />
-        <svg 
-          version="1.1"
-          baseProfile="full"
-          xmlns="http://www.w3.org/2000/svg"></svg>
+        <CountryMap strengthScores={this.state.strengthScores}  />
       </div>
     )
   }
